Fix erroneous size_t-to-int truncation.
Twice in `src/formula_function.cpp`, the result of `std::string::find`
was stored in an `int`, which was then compared for inequality with
`std::string::npos`. This was erroneous because both `std::string::npos`
and the result of `std::string::find` are of type `size_t`, so, on a
platform where `sizeof(size_t) > sizeof(int)`, the result of
`std::string::find` would be truncated to the size of an `int`, and no
mere `int` could ever equal `std::string::npos`, so the comparison would
always be true, even when it should have been false.

+++ src/formula_function.cpp
@@ -2525,7 +2525,7 @@ FUNCTION_DEF(decompress, 1, 1, "decompress(expr): Tries to decompress the given 
 		ASSERT_LOG(compressed.is_string(), "decompress takes either a compressed_data object or string.");
 		std::string s = compressed.as_string();
 		std::string key = s.length() > 10 ? s.substr(0,10) : s;
-		int n;
+		size_t n;
 		if((n = s.find(':')) != std::string::npos) {
 			key = s.substr(0, n);
 			s = s.substr(n+1);
@@ -2542,7 +2542,7 @@ FUNCTION_DEF(unencode, 1, 1, "unencode(expr) -> data_blob: Tries to unencode the
 	ASSERT_LOG(encoded.is_string(), "encoded data must be a string");
 	std::string s = encoded.as_string();
 	std::string key = s.length() > 10 ? s.substr(0,10) : s;
-	int n;
+	size_t n;
 	if((n = s.find(':')) != std::string::npos) {
 		key = s.substr(0, n);
 		s = s.substr(n+1);
